==> notebooks/01_prepare_train_data.ipynb <==
https://www.kaggle.com/competitions/amex-default-prediction/data

import gc
import os
import psutil
import pyarrow as pa
import pandas as pd

current_directory = os.getcwd()
print("Current working directory:", current_directory)

!ls -lah "C:\Users\KonuTech\zoomcamp-capstone-01\data"

# Function to get memory usage
def get_memory_usage():
    process = psutil.Process()
    mem_info = process.memory_info()
    return mem_info.rss / (1024 * 1024)  # Convert to megabytes

# Define the directory path
data_dir = os.path.join('C:\\', 'Users', 'KonuTech', 'zoomcamp-capstone-01', 'data')
data_dir

parquet_dir = os.path.join(data_dir, 'parquet_partitions')
parquet_dir

# List of file names to remove
files_to_remove = ['train_data.parquet', 'train_labels.parquet']
# Remove the files if they exist
for file_name in files_to_remove:
    file_path = os.path.join(data_dir, file_name)
    if os.path.exists(file_path):
        os.remove(file_path)

if os.path.exists(parquet_dir) and os.path.isdir(parquet_dir):
    # Directory exists, list all files in the directory
    file_list = os.listdir(parquet_dir)
    
    if file_list:
        print("List of files in the directory:")
        for filename in file_list:
            print(filename)
        
        # Delete all files in the directory
        for filename in file_list:
            file_path = os.path.join(parquet_dir, filename)
            os.remove(file_path)
            print(f"Deleted: {filename}")
    else:
        print("The directory is empty, no files to delete.")
else:
    print(f"The directory '{parquet_dir}' does not exist.")

### train_labels.csv

# Load the CSV file (train_labels.csv)
csv_file = 'train_labels.csv'
train_labels_csv_file = os.path.join(data_dir, csv_file)
train_labels = pd.read_csv(train_labels_csv_file)

# Convert the DataFrame to Parquet format
parquet_file = 'train_labels.parquet'
train_labels.to_parquet(f"{data_dir}\\{parquet_file}", index=False)

train_labels_parquet_file = os.path.join(data_dir, parquet_file)
train_labels = pd.read_parquet(train_labels_parquet_file)

train_labels.info()

train_labels.describe()

### train_data.csv

# Step 1: Read the CSV file in chunks
csv_file = 'train_data.csv'

# Step 2: Initialize an empty list to store the Parquet partition file paths
parquet_file_paths = []

chunk_size = 100000  # Adjust the chunk size as needed
i = 0  # Initialize the chunk number
cumulative_rows = 0  # Initialize the cumulative row count

if not os.path.exists(parquet_dir):
    os.makedirs(parquet_dir)


# Create a TextFileReader, which is iterable with chunks of 10,000 rows.
csv_iterator = pd.read_csv(os.path.join(data_dir, csv_file), iterator=True, chunksize=chunk_size)

parquet_file_paths = []  # Initialize the list to store Parquet partition file paths

# Iterate through the CSV file in chunks using pd.read_csv
for chunk in csv_iterator:
    # Display memory usage before reading the chunk
    before_memory = get_memory_usage()

    # Count and print the number of rows in the chunk
    num_rows = len(chunk)
    cumulative_rows += num_rows  # Accumulate the row count
    print(f"Processing chunk {i}, rows: {num_rows}, cumulative rows: {cumulative_rows}")

    # Save the chunk as a Parquet partition
    parquet_partition_file = os.path.join(parquet_dir, f'chunk_{i}.parquet')
    chunk.to_parquet(parquet_partition_file, index=False)

    # Append the Parquet partition file path to the list
    parquet_file_paths.append(parquet_partition_file)

    # Display memory usage after reading and saving the chunk
    after_memory = get_memory_usage()
    print(f"Memory usage before chunk: {before_memory:.2f} MB")
    print(f"Memory usage after chunk: {after_memory:.2f} MB")

    i += 1  # Increment the chunk number

# Concatenate the Parquet partitions into a single DataFrame
parquet_partitions = [pd.read_parquet(partition) for partition in parquet_file_paths]
df = pd.concat(parquet_partitions, ignore_index=True)

# Trigger garbage collection to clear unreferenced objects
gc.collect()

### Merge with Target

# Merge the DataFrames on the "customer_ID" column
merged_df = pd.merge(df, train_labels, on="customer_ID", how="left")

merged_df.info(verbose=True)

if os.path.exists(train_labels_parquet_file):
    os.remove(train_labels_parquet_file)
    print(f"{train_labels_parquet_file} has been removed.")
else:
    print(f"{train_labels_parquet_file} does not exist, so it wasn't removed.")

missing_values = merged_df.isnull().sum()
missing_values

# Step 4: Convert and save the combined DataFrame as a single Parquet file
combined_parquet_file = 'train_data.parquet'

merged_df.to_parquet(os.path.join(data_dir, combined_parquet_file))

# Step 5: Remove individual Parquet partitions
for partition_file in parquet_file_paths:
    os.remove(partition_file)

# Step 6: EDA on the combined Parquet file
parquet_df = pd.read_parquet(os.path.join(data_dir, combined_parquet_file))  # Read the combined Parquet file

parquet_df.info(verbose=True)

parquet_df["customer_ID"].value_counts()

parquet_df.describe()
