==> notebooks/04_get_champion_binary_classifier.ipynb <==
https://www.kaggle.com/competitions/amex-default-prediction/data

import logging
import warnings
import sys
import json
import gc
import joblib  # Import joblib for model saving
from io import StringIO
from datetime import datetime
import shutil
import pickle

import os
import pandas as pd
from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.feature_selection import VarianceThreshold, RFE
from sklearn.pipeline import Pipeline
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import f1_score
from sklearn.feature_extraction import DictVectorizer
from sklearn.impute import SimpleImputer
# from imblearn.under_sampling import RandomUnderSampler

pd.set_option('display.max_columns', None)

def calculate_correlation_with_target(features_df, target_series, variance_threshold=0.0):
    """
    Calculate the correlation of numeric columns in a features DataFrame with a target Series
    and perform Variance Threshold feature selection.
    
    Parameters:
    features_df (pd.DataFrame): The features DataFrame.
    target_series (pd.Series): The target Series.
    variance_threshold (float): Variance threshold for feature selection. Features with variance
        below this threshold will be removed. Default is 0.0 (no threshold).

    Returns:
    pd.Series: A Series containing the correlation coefficients sorted by absolute values.
    """
    # Select only numeric columns from the features DataFrame
    numeric_features = features_df.select_dtypes(include=['number'])
    
    # Calculate the correlation and sort the result by absolute values in descending order
    correlation_series = numeric_features.corrwith(target_series)
    absolute_correlation_series = correlation_series.abs()
    
    # Apply Variance Threshold to filter features
    if variance_threshold > 0.0:
        selector = VarianceThreshold(threshold=variance_threshold)
        numeric_features = selector.fit_transform(numeric_features)
        # Update correlation series to match the selected features
        correlation_series = pd.Series(selector.inverse_transform(correlation_series.values.reshape(1, -1))[0], index=numeric_features.columns)
    
    # Sort the DataFrame by absolute values
    correlation_series = correlation_series.sort_values(ascending=False)
    
    return correlation_series

# Define the directory path
data_dir = os.path.join('C:\\', 'Users', 'KonuTech', 'zoomcamp-capstone-01', 'data')
data_dir

train_data_parquet_file = 'train_data_downsampled.parquet'

# Load the training data
train_data = pd.read_parquet(os.path.join(data_dir, train_data_parquet_file))

train_data.info(verbose=True)

categorical_features = ['B_30', 'B_38', 'D_114', 'D_116', 'D_117', 'D_120', 'D_126', 'D_63', 'D_64', 'D_66', 'D_68']
categorical_features

distinct_value_counts = {}  # A dictionary to store value counts for each feature

for feature in categorical_features:
    value_counts = train_data[feature].value_counts(dropna=False)  # Count the occurrences of each distinct value
    distinct_value_counts[feature] = value_counts

print(distinct_value_counts)

result = {}  # A dictionary to store the percentage of "1" in "target" for each feature

for feature in categorical_features:
    # Calculate the percentage of "1" in "target" for the current feature
    percentages = train_data.groupby(feature)['target'].mean() * 100  # Multiply by 100 to get percentages
    
    result[feature] = percentages

# Print or use the result dictionary as needed
print(result)

### Impute median

# Specify columns to exclude from imputation
exclude_columns = ['customer_ID', 'S_2', 'target', 'D_63', 'D_64']

# Create a SimpleImputer to impute with the median
imputer = SimpleImputer(strategy='median')

# Separate the DataFrame into the columns to impute and those to exclude
columns_to_impute = train_data.columns.difference(exclude_columns)
columns_to_impute

# Fit the imputer to the data and transform the specified columns
train_data[columns_to_impute] = imputer.fit_transform(train_data[columns_to_impute])

distinct_value_counts = {}  # A dictionary to store value counts for each feature

for feature in categorical_features:
    value_counts = train_data[feature].value_counts(dropna=False)  # Count the occurrences of each distinct value
    distinct_value_counts[feature] = value_counts

print(distinct_value_counts)

# Save the fitted DictVectorizer to a file using pickle
with open('imputer.pkl', 'wb') as file:
    pickle.dump(imputer, file)

# Convert the specified columns to strings
train_data[categorical_features] = train_data[categorical_features].astype(str)

train_data.info(verbose=True)

### splitting

# Split the dataset into train and validation sets
X_train, X_val, y_train, y_val = train_test_split(train_data.iloc[:, :-1], train_data["target"], test_size=0.2, random_state=42)

# Explicitly call the garbage collector to free up memory
gc.collect()

X_train.shape

X_val.shape

y_train.shape

y_val.shape

correlation_result = calculate_correlation_with_target(X_train, y_train)

top_ten_correlations = list(correlation_result[:10].index)
top_ten_correlations

features = top_ten_correlations + categorical_features
features

### Fitting DictVectorizer

print(f"Step 1: Excluding 'customer_ID', 'S_2' and creating a DictVectorizer")

# Exclude 'customer_ID' column and create a DictVectorizer
dict_vectorizer = DictVectorizer(sparse=False)

X_train_dict = X_train[features].to_dict(orient='records')
X_val_dict = X_val[features].to_dict(orient='records')

X_train_encoded = dict_vectorizer.fit_transform(X_train_dict)
X_val_encoded = dict_vectorizer.transform(X_val_dict)

len(dict_vectorizer.get_feature_names_out())

dict_vectorizer.get_feature_names_out()

X_train_encoded[0]

### Pickling fitted DictVectorizer()

# Save the fitted DictVectorizer to a file using pickle
with open('dict_vectorizer.pkl', 'wb') as file:
    pickle.dump(dict_vectorizer, file)

gc.collect()

### Training

classifiers = [
    ('LogisticRegression', LogisticRegression(n_jobs=-1), {
        # 'classifier__max_iter': [1000],
        'classifier__solver': ['saga'],
        'classifier__C': [0.1, 1.0],
        'classifier__penalty': ['l1', 'l2'],
    }),
    ('XGBoost', xgb.XGBClassifier(eval_metric='auc', colsample_bytree=0.8, n_jobs=-1), {
        'classifier__n_estimators': [50, 100],
        'classifier__max_depth': [5, 10],
        'classifier__min_child_weight': [1, 2]  # Optimized parameter
    }),
    # ('RandomForest', RandomForestClassifier(n_jobs=-1), {
    #     'classifier__n_estimators': [200],  # [100, 200],
    #     'classifier__max_depth': [6],  # [None, 6],
    #     'classifier__min_samples_split': [5],  # [2, 5],
    #     'classifier__min_samples_leaf': [2]  # [1, 2]
    # })
]

import logging
import json
import gc
import joblib
from sklearn.feature_selection import RFE
from sklearn.impute import SimpleImputer
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import roc_auc_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import MinMaxScaler  # Import the MinMaxScaler
from datetime import datetime

# Set the log file name and log level
log_file = 'training_log.log'
log_level = logging.INFO  # You can change the log level as needed (e.g., DEBUG, INFO, WARNING, ERROR)

# Generate a timestamp
timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")

# Configure logging
logging.basicConfig(filename=log_file, level=log_level, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')

results = []

# Get the total number of logical processors
total_processors = os.cpu_count()

# Calculate the number of processors to use (80% of the available)
desired_processors = int(0.7 * total_processors)

for i, (name, classifier, params) in enumerate(classifiers, start=1):
    logging.info(f"Step {i}: Training {name} classifier")

    # Create an RFE model and a pipeline
    rfe = RFE(estimator=classifier, n_features_to_select=10)  # specify a number of features to be selected by RFE

    # Add a step to impute missing values with the median
    # imputer = SimpleImputer(strategy='median')  # You can choose a different strategy

    # Add Min-Max scaling to the pipeline
    scaler = MinMaxScaler()

    pipeline = Pipeline([
        # ('imputer', imputer),  # Add the imputation step
        ('scaler', scaler),  # Add the scaling step
        ('feature_selection', rfe),
        ('classifier', classifier)
    ])

    logging.info(f"Step {i + 1}: Performing hyperparameter tuning")

    # Perform hyperparameter tuning
    grid = GridSearchCV(pipeline, param_grid=params, cv=3, n_jobs=desired_processors, verbose=3)  # n_jobs=-1
    grid.fit(X_train_encoded, y_train)

    # Access the feature names used in the best_estimator_
    selected_feature_indices = grid.best_estimator_.named_steps['feature_selection'].get_support(indices=True)
    logging.info(f"Selected Indices of Features for {name}: {selected_feature_indices}")

    # Get selected features labels
    # feature_names = X_train_encoded.columns
    # Get the feature names from the DictVectorizer
    feature_names = dict_vectorizer.get_feature_names_out()

    # Get the selected feature names
    selected_features = [feature_names[i] for i in selected_feature_indices]

    logging.info(f"Selected Features for {name}: {selected_features}")

    # Log the output of grid.best_estimator_
    logging.info(f"Best Estimator for {name}: {grid.best_estimator_}")

    # Save the best model as a .bin file with the timestamp
    best_model = grid.best_estimator_
    model_filename = f"{name}_{timestamp}.bin"  # Include the timestamp in the filename
    joblib.dump(best_model, model_filename)

    logging.info(f"Step {i + 2}: Saving the best model for {name} as {model_filename}")

    # Explicit garbage collection
    gc.collect()

    logging.info(f"Step {i + 3}: Evaluating the best model on the validation set using Gini coefficient")

    # Evaluate the best model on the validation set using Gini coefficient
    y_prob = grid.predict_proba(X_val_encoded)  # Use predict_proba to get probabilities
    gini = 2 * roc_auc_score(y_val, y_prob[:, 1]) - 1  # Calculate the Gini coefficient

    results.append({'name': name, 'gini': gini, 'best_params': grid.best_params_})

# Create the JSON filename with the timestamp
json_file_name = f'grid_search_results_{timestamp}.json'

# Save the results to the JSON file
with open(json_file_name, 'w') as file:
    json.dump(results, file, indent=4)

# Print the logs to the terminal
log_formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
log_handler = logging.StreamHandler()
log_handler.setFormatter(log_formatter)
root_logger = logging.getLogger()
root_logger.addHandler(log_handler)

results

# Get the current working directory
current_dir = os.getcwd()
current_dir

destination_dir = os.path.join('C:\\', 'Users', 'KonuTech', 'zoomcamp-capstone-01', 'models')
destination_dir

# Specify the file extension to filter (e.g., ".bin")
file_extension = ".bin"

# List files in the current working directory
source_files = os.listdir(current_dir)
source_files

# Move files with the specified extension
for file in source_files:
    if file.endswith(file_extension):
        # Construct the source and destination paths
        source_path = os.path.join(current_dir, file)
        
        # Check if the file exists before moving
        if os.path.exists(source_path):
            destination_path = os.path.join(destination_dir, file)
            shutil.move(source_path, destination_path)
        else:
            print(f"File not found: {file}")

# Verify the move operation
destination_files = os.listdir(destination_dir)
print(f'Moved files with extension {file_extension} to destination directory: {destination_files}')

os.listdir(destination_dir)
