==> notebooks/06_score_test_data.ipynb <==
https://www.kaggle.com/competitions/amex-default-prediction/data

import os
import pyarrow as pa
import pandas as pd
import shutil

import joblib
import xgboost as xgb

import pickle
from sklearn.impute import SimpleImputer
from sklearn.feature_extraction import DictVectorizer

def score_xgb_model(model_file, input_data, columns_used):
    # Load the XGBoost model from the specified file
    xgb_model = joblib.load(model_file)
    
    # Ensure the input DataFrame contains only the specified columns
    input_data = input_data[columns_used]
    
    # Make predictions using the loaded model
    # predictions = xgb_model.predict(input_data)
    predictions = xgb_model.predict_proba(input_data)  # Use predict_proba to get probabilities

    # Extract the probability of the positive class (assuming binary classification)
    probabilities = predictions[:, 1]
    
    # Add predictions to the input data DataFrame
    # input_data['prediction'] = predictions

    # Add probabilities to the input data DataFrame
    input_data['prediction'] = probabilities
    
    return input_data

# Define the directory path
data_dir = os.path.join('C:\\', 'Users', 'KonuTech', 'zoomcamp-capstone-01', 'data')
data_dir

# Step 4: Convert and save the combined DataFrame as a single Parquet file
combined_parquet_file = 'test_data.parquet'

# Step 6: EDA on the combined Parquet file
parquet_df = pd.read_parquet(os.path.join(data_dir, combined_parquet_file))  # Read the combined Parquet file

# Shuffle the entire DataFrame randomly
parquet_df = parquet_df.sample(frac=1, random_state=42).reset_index(drop=True)

unique_clients_df = parquet_df.drop_duplicates(subset='customer_ID', keep='first')

unique_clients_df.info(verbose=True)

### Example of a Good Customer

unique_clients_df.iloc[0].to_dict()

### Example of a Bad Customer

unique_clients_df.iloc[111].to_dict()

# unique_clients_df.to_parquet(os.path.join(data_dir, "test_data_randomly_sampled.parquet"))

!ls

categorical_features = ['B_30', 'B_38', 'D_114', 'D_116', 'D_117', 'D_120', 'D_126', 'D_63', 'D_64', 'D_66', 'D_68']
categorical_features

# Step 2: Load the pickled SimpleImputer from the file
with open('imputer.pkl', 'rb') as file:
    imputer = pickle.load(file)

# Specify columns to exclude from imputation
exclude_columns = ['customer_ID', 'S_2', 'target', 'D_63', 'D_64']

# Separate the DataFrame into the columns to impute and those to exclude
columns_to_impute = unique_clients_df.columns.difference(exclude_columns)
columns_to_impute

# Fit the imputer to the data and transform the specified columns
unique_clients_df[columns_to_impute] = imputer.transform(unique_clients_df[columns_to_impute])

# Convert the specified columns to strings
unique_clients_df[categorical_features] = unique_clients_df[categorical_features].astype(str)

features = [
    'D_48',
    'D_55',
    'B_9',
    'D_58',
    'D_75',
    'B_7',
    'B_23',
    'B_16',
    'D_44',
    'B_3',
    'B_30',
    'B_38',
    'D_114',
    'D_116',
    'D_117',
    'D_120',
    'D_126',
    'D_63',
    'D_64',
    'D_66',
    'D_68'
]

### Unpickling DictVectorizer()

unique_clients_df_dict = unique_clients_df[features].to_dict(orient='records')

# Load the saved DictVectorizer from the file
with open('dict_vectorizer.pkl', 'rb') as file:
    dict_vectorizer = pickle.load(file)

# Transform the new data using the unpickled DictVectorizer
unique_clients_df_encoded = dict_vectorizer.transform(unique_clients_df_dict)

len(dict_vectorizer.get_feature_names_out())

dict_vectorizer.get_feature_names_out()

# model_file = "XGBoost_2023-11-01_17-54-58.bin"
# model_file = "XGBoost_2023-11-01_20-08-34.bin"
# model_file = "XGBoost_2023-11-04_18-06-41.bin"
model_file = "XGBoost_2023-11-05_14-54-16.bin"

models_dir = os.path.join('C:\\', 'Users', 'KonuTech', 'zoomcamp-capstone-01', 'models')
models_dir

models_path = os.path.join(models_dir, model_file)

# XGBoost_2023-11-01_17-54-58.bin :
# columns_used = ['B_16', 'B_19', 'B_20', 'B_23', 'B_3', 'B_38', 'B_4', 'B_7', 'B_9', 'D_44', 'D_48', 'D_55', 'D_58', 'D_74', 'D_75']  

# XGBoost_2023-11-01_20-08-34.bin :
# columns_used = ['B_16', 'B_19', 'B_20', 'B_23', 'B_3', 'B_38', 'B_4', 'B_7', 'B_9', 'D_44', 'D_48', 'D_55', 'D_58', 'D_74', 'D_75']  

# XGBoost_2023-11-04_18-06-41.bin :
# columns_used = ['D_48', 'D_55', 'B_9', 'D_58', 'D_75', 'D_44', 'B_7', 'B_23', 'B_16','B_3', 'D_74','B_38','B_20', 'B_4', 'B_19', 'B_1', 'B_37', 'B_8', 'B_22', 'B_11', 'B_30', 'S_7', 'R_1', 'S_3', 'D_70']

# XGBoost_.bin :
# columns_used = ['B_16', 'B_23', 'B_3', 'B_30=0.0', 'B_30=1.0', 'B_30=2.0', 'B_30=nan', 'B_38=1.0', 'B_38=2.0', 'B_38=3.0', 'B_38=4.0', 'B_38=5.0', 'B_38=6.0', 'B_38=7.0', 'B_38=nan', 'B_7', 'B_9', 'D_114=0.0', 'D_114=1.0', 'D_114=nan', 'D_116=0.0', 'D_116=1.0', 'D_116=nan', 'D_117=-1.0', 'D_117=1.0', 'D_117=2.0', 'D_117=3.0', 'D_117=4.0', 'D_117=5.0', 'D_117=6.0', 'D_117=nan', 'D_120=0.0', 'D_120=1.0', 'D_120=nan', 'D_126=-1.0', 'D_126=0.0', 'D_126=1.0', 'D_126=nan', 'D_44', 'D_48', 'D_55', 'D_58', 'D_63=CL', 'D_63=CO', 'D_63=CR', 'D_63=XL', 'D_63=XM', 'D_63=XZ', 'D_64=-1', 'D_64=None', 'D_64=O', 'D_64=R', 'D_64=U', 'D_66=0.0', 'D_66=1.0', 'D_66=nan', 'D_68=0.0', 'D_68=1.0', 'D_68=2.0', 'D_68=3.0', 'D_68=4.0', 'D_68=5.0', 'D_68=6.0', 'D_68=nan', 'D_75']
columns_used = ['B_16', 'B_23', 'B_3', 'B_30=0.0', 'B_30=1.0', 'B_30=2.0',
       'B_38=1.0', 'B_38=2.0', 'B_38=3.0', 'B_38=4.0', 'B_38=5.0',
       'B_38=6.0', 'B_38=7.0', 'B_7', 'B_9', 'D_114=0.0', 'D_114=1.0',
       'D_116=0.0', 'D_116=1.0', 'D_117=-1.0', 'D_117=1.0', 'D_117=2.0',
       'D_117=3.0', 'D_117=4.0', 'D_117=5.0', 'D_117=6.0', 'D_120=0.0',
       'D_120=1.0', 'D_126=-1.0', 'D_126=0.0', 'D_126=1.0', 'D_44',
       'D_48', 'D_55', 'D_58', 'D_63=CL', 'D_63=CO', 'D_63=CR', 'D_63=XL',
       'D_63=XM', 'D_63=XZ', 'D_64=-1', 'D_64=None', 'D_64=O', 'D_64=R',
       'D_64=U', 'D_66=0.0', 'D_66=1.0', 'D_68=0.0', 'D_68=1.0',
       'D_68=2.0', 'D_68=3.0', 'D_68=4.0', 'D_68=5.0', 'D_68=6.0', 'D_75']

len(columns_used)

# Convert the NumPy array to a DataFrame
unique_clients_df_encoded_df = pd.DataFrame(unique_clients_df_encoded, columns=columns_used)  # Replace 'your_column_names' with the actual column names

for column_name in columns_used:
    if column_name not in unique_clients_df_encoded_df.columns:
        print(f"Column '{column_name}' not found in the input_data DataFrame.")

### Scoring of a data

df_scored = score_xgb_model(
    model_file=models_path,
    input_data=unique_clients_df_encoded_df,
    columns_used=columns_used
)

df_scored.info(verbose=True)

df_scored

unique_clients_df.shape

unique_clients_df['customer_ID']

df_scored.shape

df_scored["prediction"]

# Assuming you have two Series: unique_clients_df["customer_ID"] and df_scored["prediction"]
customer_ID_series = unique_clients_df["customer_ID"].reset_index(drop=True)
prediction_series = df_scored["prediction"].reset_index(drop=True)

# Concatenate the two Series by row indices
merged_df = pd.concat([customer_ID_series, prediction_series], axis=1)

# Optional: Rename the columns if needed
merged_df.columns = ["customer_ID", "prediction"]

merged_df.info()

merged_df

# merged_df.to_csv("submission_2023-11-01_17-54-58.csv", index=False)
# merged_df.to_csv("submission_2023-11-01_20-08-34.csv", index=False)
# merged_df.to_csv("submission_2023-11-04_18-06-41.csv", index=False)
merged_df.to_csv("submission_2023-11-05_14-54-16.bin.csv", index=False)

# Get the current working directory
current_dir = os.getcwd()
current_dir

destination_dir = os.path.join('C:\\', 'Users', 'KonuTech', 'zoomcamp-capstone-01', 'data')
destination_dir

# Specify the file extension to filter (e.g., ".bin")
file_extension = ".csv"

# List files in the current working directory
source_files = os.listdir(current_dir)
source_files

# Move files with the specified extension
for file in source_files:
    if file.endswith(file_extension):
        # Construct the source and destination paths
        source_path = os.path.join(current_dir, file)
        
        # Check if the file exists before moving
        if os.path.exists(source_path):
            destination_path = os.path.join(destination_dir, file)
            shutil.move(source_path, destination_path)
        else:
            print(f"File not found: {file}")

# Verify the move operation
destination_files = os.listdir(destination_dir)
print(f'Moved files with extension {file_extension} to destination directory: {destination_files}')

os.listdir(destination_dir)

### Example of a Good customer

df_scored.iloc[0].to_dict()

### Example of a Bad customer

df_scored.iloc[111].to_dict()
